Changed routes to PUT operations and removed ID from the request path, since we can take it from the body of the request
Since we give in the PUT operation, and also in POST, I followed the example from there, and I removed the path piece containing the ID. For example, from now we can do an Update just by accesing /students instead of /students/ID and do a PUT using a json representing a Student. The service will then search for the ID given in the request body, instead of searching the ID given in the link. It is a bit cleaner this way